+++ kirlent/slides.py
@@ -13,62 +13,61 @@
 # You should have received a copy of the GNU General Public License
 # along with Kırlent.  If not, see <http://www.gnu.org/licenses/>.
 
+from io import StringIO
 from pathlib import Path
+from typing import Dict
 
 from invoke import Context, task
 
-from .utils import (
-    MKDIR,
-    collect_assets,
-    relative_path,
-    relativize_paths,
-    up_to_date
-)
+from .utils import MKDIR, collect_assets, relative_path, up_to_date
 
 
-BUILD_SLIDES: str = "%(builder)s %(options)s %(in)s %(out)s"
+BUILDERS: Dict[str, str] = {
+    "simple": "kirlent2slides",
+    "impressjs": "kirlent2impressjs",
+    "revealjs": "kirlent2revealjs",
+}
+
+BUILD_SLIDES: str = "%(builder)s %(options)s %(in)s"
 
 
-def slides(c: Context, unit: str, *, framework: str, lang: str = "*") -> None:
-    unit_path = Path(unit)
-    slug = relative_path(unit_path, Path(c.contents))
-    output_path = Path(c.output, slug)
-
-    for src in unit_path.glob(f"slides.{lang}.rst"):
-        language = src.name.split(".")[1]
-        target_name = f"slides-{framework}.{language}.html"
-        target = output_path / target_name
-        raw_target = output_path / f".{target_name}"
-        if not up_to_date(raw_target, [src]):
-            if not raw_target.parent.exists():
-                c.run(MKDIR % {"dir": relative_path(raw_target.parent)})
-            options = c.slides.options.base + c.slides.options.get(framework, [])  # noqa: E501
-            cli_options = " ".join(options) % {
-                "lang": language,
-                "size": c.slides.size,
-            }
-            format = "slides" if framework == "simple" else framework
-            c.run(BUILD_SLIDES % {
-                "builder": f"kirlent2{format}",
+def slides(c: Context, src: Path, output: Path, *, framework: str) -> None:
+    suffixes = "".join(src.suffixes)
+    dst_name = src.name.replace(suffixes, f"-{framework}{suffixes}")
+    dst = Path(output, dst_name).with_suffix(".html")
+    if not up_to_date(dst, [src]):
+        if not dst.parent.exists():
+            c.run(MKDIR % {"dir": relative_path(dst.parent)})
+        lang = src.name.split(".")[1]
+        options = c.slides.options.base + c.slides.options.get(framework, [])
+        cli_options = " ".join(options) % {
+            "lang": lang,
+            "size": c.slides.size,
+        }
+        out = StringIO()
+        c.run(
+            BUILD_SLIDES % {
+                "builder": BUILDERS[framework],
                 "options": cli_options,
                 "in": relative_path(src),
-                "out": relative_path(raw_target),
-            })
-            relativize_paths(raw_target, src.parent)
-
-        collect_assets(c, raw_target, target)
-
-
-@task
-def simple(c, unit, lang="*"):
-    slides(c, unit, lang=lang, framework="simple")
+            },
+            out_stream=out,
+        )
+        text = collect_assets(c, out.getvalue(), base=output, wrt=src.parent)
+        dst.write_text(text)
+    c.config[f"{framework}:output"] = str(dst)
 
 
 @task
-def impressjs(c, unit, lang="*"):
-    slides(c, unit, lang=lang, framework="impressjs")
+def simple(c, src, output):
+    slides(c, Path(src), Path(output), framework="simple")
 
 
 @task
-def revealjs(c, unit, lang="*"):
-    slides(c, unit, lang=lang, framework="revealjs")
+def impressjs(c, src, output):
+    slides(c, Path(src), Path(output), framework="impressjs")
+
+
+@task
+def revealjs(c, src, output):
+    slides(c, Path(src), Path(output), framework="revealjs")